re-refeactor game loop
